requirements.txt
speechrecognition
gtts
pygame
colorama
openai
python-dotenv
